Fix: remove a typo that is only exposed for some less common OSes
#### Brief overview of PR changes/additions
Fix a typo that is only exposed for some OSes (not Windows or GNU/Linux and
probably not MacOS either) and also reorder some tests so that, for
example, each of Open/Free/NetBSD get reported as their respective ones
rather than all of them being identified as a BSD4.4 OS.

#### Motivation for adding to Mudlet
When I returned to a FreeBSD OS recently I tried to build the current
`development` branch code and found that, due to some faulty ordering of
some OS detection I had enabled some BSD4 OS type code which had a typo
making the code not compile.

#### Other info (issues closed, discussion etc)
There is a similar piece of code in `TLuaInterprete::getOS(...)` which
I had done previously and which had a logic behind the order of the tests,
I referred to that when reordering the ones in this bit of code.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>