Fix: correct some inconsistencies in embedded Build version in resources
#### Brief overview of PR changes/additions

Change a CMake project file so that the Git SHA1 is always included in
the "Application Build" value that gets stored in the app_build.txt file.
The qmake project file did do that, but if a value was in the
`MUDLET_VERSION_BUILD` environmental variable it was not being included in
the "Version" text shown within Mudlet with that build system.

Also changes the mix of `BUILD_TEST` and `BUILD` in the qmake project file
back so that they are all `BUILD`.

Also apply/change some prefixes to indicate class membership and
const-ness - some of which have been changed

Now closes the app-build.txt file after it has been read from the
embedded resources. Also trim off the trailing white-space (end of line)
from the text read from that. Also, now only read the file once, and pass
the detail into the `mudlet` class constructor, which has been revised so
that some key `bool` and `QString` values have been moved back into the
class as `const` members. This avoids some awkward to do `static const`
variable initialisation which could not trim white space from the text.

#### Motivation for adding to Mudlet
I had found that setting a `MUDLET_VERSION_BUILD` environmental variable
in my builds was no longer showing up in the displayed version information
for qmake type builds.

Furthermore due to a previous incomplete renaming of the Mudlet `BUILD`
qmake variable to `BUILD_TEST` the Application icon was not being correctly
set on Windows builds.


The naming of some of the `mudlet` class members no longer matched up to
the prefixing {`c` = `const`, `s` = `static`, `m` = member`} that has been
used. This PR tries to sort out that, especially as some have been changed
by this PR.


#### Other info (issues closed, discussion etc)
Apply some prefixes to indicate class membership and const-ness - some
of which have been changed:
* `(bool) mudlet::releaseVersion`
                                ==> `(const bool) mudlet::cmReleaseVersion`
* `(bool) mudlet::publicTestVersion`
                             ==> `(const bool) mudlet::cmPublicTestVersion`
* `(bool) mudlet::developmentVersion`
                            ==> `(const bool) mudlet::cmDevelopmentVersion`
* `(QString) mudlet::scmVersion` ==> `(const QString) mudlet::cmVersion`
* `(QString) mudlet::mAppBuild` ==> `(const QString) mudlet::cmAppBuild`
* `(static void) mudlet::setNetworkRequestDefaults(const QUrl&,
                                                          QNetworkRequest&)`
     ==> `(void) mudlet::setNetworkRequestDefaults(const QUrl&,
                                                          QNetworkRequest&)`

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>