Infrastructure: split TLuaInterpreter.cpp's TTS files out separately (#7135)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Split TLuaInterpreter.cpp into a domain-specific TTS functions file.
#### Motivation for adding to Mudlet
A 18k line C++ files makes a lot of tools cry.
#### Other info (issues closed, discussion etc)
I've purposefully put the new files into a similar but random-ish
location in mudlet.pro and CMakeLists.txt to reduce the changes of
conflicts when merging the PRs. We can fix the order in a single PR
afterwards.