Fix: remove double SHA1 from CI builds
It was breaking the PTB builds because their version identification was
malformed with two copies of the development branch git SHA1 for PTB
(and presumably) Release builds. Testing and PR builds had the HEAD SHA1
for the current branch applied first followed by that of the development
branch commit that the branch was rebased onto. Given the values concerned
the first one (which is now generated internally by the QMake/CMake build
process) seems to be the one that we want to keep now it is done there
(so that it can be embedded in a "resource" file rather than passed to
each object file that gets compiled) this PR attempts to fix things by
no longer appending the Git SHA1 to the "MUDLET_VERSION_BUILD"
environmental variable.

Note: there is some cruft in the `CI/travis.set-build-info.sh` file
relating to builds on the Travis platform - but seen they abandoned
supporting FOSS projects a few years back there is no need for that baggage
however, to keep things simple I have updated that fragment even though it
is ripe for excision from there.

Hopefully this will close #7158 and #7159.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>