Revise: use PR commit SHA instead of commit of development used for Windows
The wrong commit was being used.

However it is possible that the qmake processing to get the SHA to use is
not behaving as expected and is also getting the development branch head
into which the PR is merged to use. This will need further investigation
and will show up as the wrong SHA1 embedded in the resource files for
Windows builds.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>