Fix: identify correct SHA1 to use on AppVeyor for Windows
There we need to identify the grandparent of the HEAD commit we are
informed about. This needs to go into the qmake project file (since we now
embed the SHA1 into the resources file) as well as the (Windows)
"set-build-info" script. This issue was already handled for the Linux/MacOS
CI/CB processes in the GitHub actions by the line:
```sh
BUILD_COMMIT=$(git rev-parse --short "${GITHUB_SHA}^2")
```
in the corresponding "set-build-info" file. We are using `~2` rather than
`^2` as because (AFAICT) the former always refers to the ancestor in the
"working" branch whereas the later does only when there are merges in the
neighbourhood - and since the `git rev-parse` is always run it will not
produce error messages about a non-existent commit reference when there
isn't one.

This commit does not address the CMake case because it is not used/usable
in that environment for the moment - but if this works I will port the
same process to that project file.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>