Fix: correct logic error that meant dodgy codepoints were always reported
#### Brief overview of PR changes/additions
Change some boolean ORs to ANDs - making the reporting of problematic code-
points much less spammy all around.

#### Motivation for adding to Mudlet
There is a setting so that when a codepoint is encountered by the `TTextEdit`
class where there is a code-point whose width in terms of narrow or wide
character is not determinable (or it does not have one) a debug message
is printed to the OS command line - either just at the end of a session
or for the first time this happens per codepoint per session. Unfortunately
a logic error on my part in #3832 meant that every single time such a code-
point was encountered a new message was displayed.

#### Other info (issues closed, discussion etc)
Thanks to humera1295 (Discord user) for bringing this to my attention.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>