Improve: add support for a codec to support a revised Medievia custom font (#7161)
#### Brief overview of PR changes/additions
This PR adds a new custom codec that aims to map the Extended ASCII
character codes that **Mudievia** uses for use with its [custom `.ttf`
font](http://www.medievia.com/fonts/MedSansMono.ttf).

#### Motivation for adding to Mudlet
Unfortunately the WINDOWS-1252 encoding that Medievia appears to use
includes some character codes that are NOT defined for that encoding
which Mudlet, not unreasonably rejects and replaces with the
"Replacement character" glyph. Also, some of the codes it uses when put
through that WINDOWS-1252 encoding map to C1 control characters which
are also not displayable because Mudlet's use of RidiculousFish's
`widechar_width.h` function decodes them as not visible characters. The
solution is to map those glyphs that can be positioned into their
"standard" codepoint in the font and put in "references" to them in the
original places in the font, for the other "custom" glyphs they can be
positioned in a block of a Private Use Area where Mudlet's new custom
codec can access them without losing them in the text painting part of
Mudlet.

#### Other info (issues closed, discussion etc)
To test this PR the revised font will also be needed, I am preparing
that in a separate repository of my own:
https://github.com/SlySven/Medievia_font

At this point the font has been updated in a manner to work with
Mudlet after this PR, using version number of 3.0.0 or higher.

It is possible that the "Show Control Characters" as "OEM Font" (CP-437)
characters option will need to be selected as well as Medievia does
appear to use those as well. Unfortunately the character `0x04` Diamond
will currently get eaten by Mudlet as a result of #3526 which included
code to hide an unexpected one of these when MXP is turned on or off in
Achaea, apparently.

Also:
I spotted a coding error in nearby other codecs that we were checking
for an isolated `(bool) QChar::isHighSurrogate()` when it should have
been a `(bool) QChar::isLowSurrogate()` call instead!

---------

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>
Co-authored-by: Vadim Peretokin <vperetokin@hey.com>