Improve: simplify some MMCPServer methods
They were returning values that we don't use and the remainder could be
written more succinctly. It was also possible to reduce the warnings that
a code-analyser in Qt Creator now produces about missing direct `#include`s
- which were popping up (for all sorts of things) but in this case for the
explicit use of `QPair` which we can avoid by the use of `auto` to allow
the compiler to deduce from the function that is called to assign value(s)
to a variable.

I found similar things in the dlgIRC class and have a separate PR to do the
same changes there.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>