Improve: add controlCharacterHandling to Lua get- & set-config
#### Brief overview of PR changes/additions
Extend `TLuaInterpreter::setConfig(...)` and `TLuaInterpreter::getConfig(...)`
for an additional case: `controlCharacterHandling` and also add that "key"
to the list in the external `./mudlet-lua/lua/Other.lua` script file so
the no argument `getConfig()` case includes it as well.

#### Motivation for adding to Mudlet
Medievia MUD will want this to enable a package to configure this to `"oem"`
(out of the other `"asis"` and `"picture"` options).

#### Other info (issues closed, discussion etc)
Also rename a couple of variables in the section of `"caretShortcut"` for
the `setConfig(...)` case as it was using `key`/`keys` for things that are
`value`/`values`!

Also improve a couple of other setConfig cases (`caretShortcut` and
`commandLineHistorySaveSize`) which take one of a number of values so that
entering an invalid string produces a `nil` + error message rather than
throwing an error - as is our practice for other run-time value (rather
than type) problems.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>