Improve: add named captures to Aliases (#7175)
#### Brief overview of PR changes/additions
This has been done by carefully comparing the matching code from the
PCRE trigger item. Some variables have been renamed so they match up
better.

#### Motivation for adding to Mudlet
This has been a long-standing but unlisted wish-list item but someone
actually raised an issue for it! :grinning:

Fortunately much of the needed code was already in place, just the bit
that injected the named group results into the Lua sub-system was
missing (the call to `TLuaInterpreter::setCaptureNameGroups(nameGroups,
namePositions)`)!

#### Other info (issues closed, discussion etc)
It does actually make writing aliases a bit easier because it is easier
to construct an alias using `matches.target` say rather than
`matches[4]` if `?<target>` has been inserted at the start of the
relevant capture group. It also makes extending/modifying an alias
easier as there is no need to juggle indexes in the Lua script that
uses, say, `matches.target` compared to one that uses `matches[4]`.

This should close #7171.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>