Infrastructure: reposition const before type specifier
#### Brief overview of PR changes/additions
This PR tries to put all `const`s before the type (class).

#### Motivation for adding to Mudlet
There is a mix of positioning of `const`s where it is used to indicate that
a variable is not to be modified by program code, we tend to put it before
the type but a prior PR (which looks to have been done with an automated
tool) has resulted in a mix of cases some with the const adjacent to the
variable. This lack of consistency can be confusing.

#### Other info (issues closed, discussion etc)
It looks like this came about in #6843.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>