Fix: handle some pathological cases
If the peer held in the `(QList<MMCPClient*>) MMCPServer::mPeersList`
should happen to be a `nullptr`, we do not want to try and dereference it!

Also, convert the above to `(QList<QPointer<MMCPClient>>)` so that a peer
(`MMCPClient`) getting `delete`ed (destroyed) will not cause things to
blow up - because the `QPointer` will automatically revert to a `nullptr`
when the pointed to item is destroyed.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>