Fix: clean up some rough edges
I found that I could get a seg. fault if a peer sent a message while the
profile (or maybe the whole application) was terminating!

We needed to have non-white-space characters on some strings otherwise they
got trimmed too much when painted onto the main console. This means that
an "Id" was needed on the lists of clients. The format/presentation of the
Id number and Port numbers are also tweaked.

The local display of emotes was also tweaked to show the same (hopefully)
in English as peers see as well as information that they sent an emote.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>