Infrastructure: refactor and simplify profile closing (#7203)
#### Brief overview of PR changes/additions
An attempt to improve the process of closing multiple profiles,
including handling the case where a closure of the whole Mudlet
application is cancelled part way through by the end-user selecting the
"Cancel" option when asked if they want to save a profile or not.

#### Motivation for adding to Mudlet
When multi-playing carrying out the above action (cancelling part way
through) can leave widgets and dialogues for profiles open even though
the underlying `Host` instance has gone away. This can easily lead to
fatal segmentation faults when those widgets or dialogues try to
reference something that is no longer present. Additionally the code
prior to this PR had multiple execution paths when "closing" a profile
leading to some of them not doing all the things they needed to - hence
the issues just listed. This PR tries to rationalise the execution path
and reduce the amount of duplication so that it is easier to follow and
amend in the future.

#### Other info (issues closed, discussion etc)
A `closeEvent(QEvent*)` for the `TMainConsole` class has been created to
override the `TConsole` one as it is the only console that has to be
concerned with saving the profile.

The way the code was structured meant that the destruction of
`TDockWidget`s would sometimes hit a
`"TConsole::closeEvent(QCloseEvent*) INFO - closing a UserWindow but the
TDockWidget pointer was not found to be removed...";` message - this was
because the relevant `TDockWidget` had already been removed from the
`(QMap<QString, TDockWidget*>) TMainConsole::mDockWidgetMap` before the
`TConsole::closeEvent(...)` was executed.

The `(bool) Host::mIsGoingDown` flag was unnecessary and seemed to
duplicate the `(bool) Host::mIsClosingDown` one - OTOH I did need a new
`(bool) Host::mForcedClose` flag to signal that the main application was
being forcibly closed and that everything should be saved whether the
user had set the "auto-save on exit option" or not. The `(bool)
TConsole::mUserAgreedToCloseConsole` flag has been renamed and moved to
be `(bool) TMainConsole::mEnableClose` instead.

The actions to close down widgets and dialogues associated with a
profile have been moved to the new `(void) Host::closeChildren` method
which is called by ~~both the `(void) Host::forceClose()` method - which
cannot be aborted and~~ the `(bool) Host::requestClose()` which can be
aborted, and returns `false` if that *has* happened. ***Edit: the
forceClose() method now just sets a flag which forces the requestClose()
to not ask questions and just save things...***

The `HostManager::deleteHost(const QString&)` return value was never
used so I changed it to be a `void` method. Note that this is the method
that is actually responsible for destroying the `Host` instance for each
profile.

The `(static void) TDebug::removeHost(Host*)` method needs the name of
the profile that is "going away" but because it is called from the
`Host` destructor I ran into a copy of seg. faults from:
* calling `Host::getName()` from there
* calling `TBuffer::log(...)` indirectly from there because that
dereferences `Host::mpConsole` to check if the `TBuffer` is the one for
the main console and there wasn't a valid `Host` at that point.

The process of iterating through all the open profiles to close them
down in `(void) mudlet::closeEvent(QCloseEvent*)` is more tricky than it
seems because the `for (auto pHost : mHostManager)` loop blows up if
`mHostManager` is modified in anyway during the iteration. Instead I
found I had to note down the profiles that have been closed and only
destroy them after the "Do you wanna save this profile?" loop has been
gone through for as many as get closed.

Because the package exporter widget is parented to the widget that is
effectively the `QMainWindow` for the mudlet instance rather than the
`TMainConsole` for it's profile, it was not getting closed/destroyed
should it have been open when the profile it was associated with get
closed but the Mudlet application stay running. As the `Host` class is
NOT derived from a `QWidget` the dialogue could not be made a child of
it; instead I've arranged for the `QObject::destroyed` signal of the
`Host` to cancel and close it instead.

---------

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>
Co-authored-by: Kebap <kebap_spam@gmx.net>
Co-authored-by: Vadim Peretokin <vperetokin@hey.com>
Co-authored-by: Mike Conley <sousesider@gmail.com>