Improve: add faint support and separate bold from colours (#7010)
#### Brief overview of PR changes/additions
This disconnects the link that Mudlet had between ANSI colour numbers 8
to 15 (in the 16 and 256 colour ranges) from emboldening (increasing the
"weight") of the Font used.

It also adds support for the ANSI Escape SGR sequence "2" option for
"faint" (a.k.a. "dim") which provides two additional "weights" that can
be combined with the bold effect.

So as to enable Mudlet to work with 16-Colour MUDs that still use the BOLD
attribute to select the second set of 8 Colours a "BOLD is Bright" option is
now provided - so as to emulate prior Mudlet behaviour it defaults to being
enabled but in most cases users will want to disable it so that the Bold
effect is used only to change the way the font is presented and possibly
be combined with the Faint effect - provided that the font in use has enough
weights.

#### Motivation for adding to Mudlet
I have often wondered about why we have been apply the bold font effect
to characters that use the second eight of the ANSI 16 / 256 colour
indexes - and, since it dates back to early Heiko era code (commit 
3e127f554bb23f8e86fbe64c560bccfc7bd096d0 ) it is hard to see as there is
no documentation. My best *guess* is that it was intended to emphasis
that the colour was from the second eight of the sixteen which for Muds
that use that many colours are the brighter ones (with the RGB
components at "full-strength") compared to the first eight which are
less bright/intense.

#### Other info (issues closed, discussion etc)
This should help to close #3693.

I am considering providing (in a separate PR) a button on the display
preferences for the colours of the main console that will "Reset to
8-color defaults" and to rename the existing one "Reset to 16/256 color
defaults" which will swap (all but the "blacks") the first 8 with the
second 8 so that MUDs that only use 8 colours will automagically show
the "full-strength" colours for the eight that they do use.

To demonstrate all four font weights (and to check that it works with
the selected font) the following sample code could be used:

```lua
feedTriggers("\027[2mFaint \027[22mNormal \027[1;2mDemi-bold \027[22;1mBold\027[22m\n")
```
---------

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>