Fix: correct an error I made in PR7160
#### Brief overview of PR changes/additions
I made an error in #7160 as I thought that something that looked like a
function call returned a value that could be assigned to something with a
`set(...)` but CMake doesn't work like that, instead one has to provide a
variable to recieve the value as the last argument to a `string(...)`
function. :roll_eyes:

#### Motivation for adding to Mudlet
Fix an error that causes cruft to appear, at least, in the version number
on the first tab of the "About Mudlet" dialogue.

#### Other info (issues closed, discussion etc)
This should close #7220.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>