Infrastructure: move to a MSYS2 & Mingw-w64 Appveyor Windows CI build [skip actions]
#### Brief overview of PR changes/additions
This is intended to develop a pure MSYS2 and Mingw-w64 based Windows CI
build system on AppVeyor using a triplet of bash scripts I have previously
put together.

#### Motivation for adding to Mudlet
Our current Window CI uses some components of the original Mingw project
and those are becoming unusable (it appears as though one of the major
mirrors is in Ukraine and the other in Japan) and we are consistently
getting things time out in the hour allowed on Appveyor)

#### Other info (issues closed, discussion etc)
The new shell scripts that this PR provides do seem to have a
chicken-and-egg situation in that the first `setup-windows-sdk.sh` is
contained in the git repository but can be used to install the repository
and setup the required things locally. As such we can/should direct
interested parties to get the raw file independently from our GH repository
and to execute it in a MINGW64 shell (and then a MINGW32 one if 32-bits
builds are also wanted).

Once that has been done once it can be re-run to get extra things that
might be wanted.

This is being developed independently of John McKisson's work as it is
intended to be a more thorough system that can work on both Qt 5 and 6 and
also for both 32 and 64 bits and debug as well as release type builds;
although 32 bits is now only possible in a Qt 5 environment.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>