Improve: speed up opening Mudlet (#7215)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Speed up opening Mudlet as well as opening additional profiles
#### Motivation for adding to Mudlet
A speedier Mudlet makes happier players
#### Other info (issues closed, discussion etc)
We were always creating the Mudlet editor dialog (triggers/aliases/etc)
when creating a new profile - defer the creation until we actually need
to open it for the first time.

This also meant that `pEditor->fillout_form();` was getting called
twice, once at creation and another time when we actually needed it -
and this is a somewhat expensive method. Thus in the end we are not just
deferring the cost, but cutting it significantly as well.