Infrastructure: additional fixes to dblsqd registration for windows  (#7234)
The Changelog creation was attempting to use Lua 5.4 and hence did not
locate the installed version 5.1 argparse luarock. Changed the lua call
to lua5.1 to explicitly invoke the lua5.1 interpreter. Additionally
added the required lunajson rock for Changelog creation.

The dblsql call was failing possibly due to the version string not being
properly lowercased, so trying again with it all lowercase instead of
just the commit sha.

Also noticed that I had set the file_cat to 3 but it should have been 0
for Windows according to the appveyor script.