Fix: fully undo bodges to versioning
Undoes the parts of:
https://github.com/Mudlet/Mudlet/commit/fa8546e254684321c429e5bf67b55902abc38a0f
that bodged the QMake `BUILD` variable so that it **does** consider the
value set in the `MUDLET_VERSION_BUILD` *environment* variable.

This is important for me, at least, as I use the environmental variable
to tag my local builds - and others producing patched versions of Mudlet
might/ought to do the same. It is also important as this environmental
variable is also used during the CI processess to control what is built by
them.

Importantly the CMake and QMake meta-build processes (they make the actual
Makefiles that are used to build things) are normally set to ensure the
`BUILD` variable is **never** the empty string as that is what is used to
trigger the special casing for proper "Release" builds - though that does
mean that a Mudlet Core Dev does need to adjust the top-level
`CMakeLists.txt` and the `/src/mudlet.pro` files to get the `BUILD`
variable to **be** an empty string for that type of build - and to restore
the "normal" arrangement afterwards.

Actually I have changed the QMake internal variable from `BUILD` to
`APP_BUILD` so that it is the same as the one in the CMake file.

Also up the minimum CMake version to 3.5 to silence warnings in CMake 3.29
that support for older versions will be dropped soon.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>