Fix: stop bold and faint from affecting game text if BoldIsBright option used
#### Brief overview of PR changes/additions
Gate the action of the Bold and Faint SGR codes so that they only have
the presentation effects on the text font if the BoldIsBright option is
NOT set.

#### Motivation for adding to Mudlet
It has been noted that recent changes made by #7010 changed the way that
the Set Graphics Rendiion for the BOLD effect (value `1`) now aways
emboldens text when it is seen instead of not doing so when it is
associated with a basic eight fore ground colour code (values `30` to
`37`).

#### Other info (issues closed, discussion etc)
This could close #7256.

A rethink has caused me to consider whether I can make it so that bold (and
faint) can still be used if "Bold is Bright" is enabled by only considering
the `1` code to be used to select the brighter colour if it is within the
same `<SGP>` sequence. Otherwise, if it is separate, then it still does
the bold/faint effect. It may be that Heiko's original code effectively did
that!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>