Fix: stop bold and faint from affecting game text if BoldIsBright option used (alternative)
#### Brief overview of PR changes/additions
Provide an option so that the action of the Bold and Faint SGR codes do
NOT (by default) have the presentation effects on the text font if they
appear within the **same** SGR sequence that sets one the basic 8-colors
for a foreground (with a priority over) or a background color. This is so
that MUDs that use the Bold option to select the second eight colors in a
16-color do not also set the text format at the same time. The bold (and
faint) options can still be set if done in a separate SGR sequence. If
the "Bold is bright" option is not enabled (which *is* preferable for MUDs
that use all other color mods then this particular means of selecting 16
colors) then the bold and faint effects will have their intended (as far
as the EMCA standard suggests) whenever and whereever they appear in an SGR
sequence.

#### Motivation for adding to Mudlet
It has been noted that recent changes made by #7010 changed the way that
the Set Graphics Rendiion for the BOLD effect (value `1`) now aways
emboldens text when it is seen instead of not doing so when it is
associated with a basic eight fore ground colour code (values `30` to
`37`).

#### Other info (issues closed, discussion etc)
This could close #7256.

In comparision with #7277 this PR is a bit more complex but it also allows
for 16 colors to be used for the BACKground color in those MUDs that use
such a range of colors (by setting a background color WITH the bold
attribute in a separate sequence to the foreground color).

I haven't yet tested every combination of values in SGR sequences with this
PR I hope however that testers will report any that do not work as
expected!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>