Fix: get background colours working correctly for BoldIsBright
I'd left some lines of code in place that produced the wrong effect.
This was clearing the flag that we had an basic 8-color background when we
encountered an 8-color foreground colour - the proritising the latter over
the former needs to be done when we have parsed all the values in an SGR
sequence **not** when we see an appropriate foreground value.

I also had to set the background colour when a 8 color foreground was
"shifted" to a bright one.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>