Infrastructure: remove AppVeyor and other unused CI/CB files
Since we are abandoning the use of AppVeyor and putting all our CI/CB eggs
in a Git Hub Actions basket we should remove the files that are not now
needed for it. There are also some other files that are not useful anymore
and yet some others that should be listed in the Qmake project file so they
can be searched/seen in Qt Creator if used for qmake builds.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>