Fix: allow BOLD to always be for 16 color selection & never a font effect
#### Brief overview of PR changes/additions
This suppliments PR #7280 to track the status of having an 8 color
foreground or background color and a bold attribute when the BoldIsBright
option (which has been extended to a third "always" case) is set to that
"alway" on mode.

#### Motivation for adding to Mudlet
It is hoped that this will overcome the regressions that some users are
unhappy about with the **4.18.x** releases.

#### Other info (issues closed, discussion etc)
This should allow users to continue to use Mudlet on MUD servers that use
the `<SGR>1m`(BOLD) code only for selecting from 16 colors and do not care
about making text **bold**.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>