Fix: get BUILD and GIT_SHA1 values back into qmake driven builds
#### Brief overview of PR changes/additions
Restores the inclusion of any MUDLET_VERSION_BUILD environmental variable
and the Git SHA1 in the version information for anything other than release
builds. The problem was caused by the incorrect use of a single `$` instead
of the double `$$` needed to get the values of QMake variables during the
non-build (first) pass through the project file.

#### Motivation for adding to Mudlet
It is breaking PT builds as it is mangling the version information.

#### Other info (issues closed, discussion etc)
This should close issue #7305.

This PR also adds the name of some existing files to the QMake project file
so they show up in Qt Creator.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>